Updated linux install script

+++ install-linux.sh
@@ -2,6 +2,33 @@
 
 MINECRAFT_FOLDER=~/.minecraft
 
+uuid()
+{
+    local N B C='89ab'
+
+    for (( N=0; N < 16; ++N ))
+    do
+        B=$(( $RANDOM%256 ))
+
+        case $N in
+            6)
+                printf '4%x' $(( B%16 ))
+                ;;
+            8)
+                printf '%c%x' ${C:$RANDOM%${#C}:1} $(( B%16 ))
+                ;;
+            3 | 5 | 7 | 9)
+                printf '%02x-' $B
+                ;;
+            *)
+                printf '%02x' $B
+                ;;
+        esac
+    done
+
+    echo
+}
+
 #install build tools
 sudo apt-get update
 sudo apt-get install -y python libpng-dev libjpeg-dev libboost-iostreams-dev \
@@ -16,11 +43,20 @@ make
 cd ../../
 
 #install minecraft textures
-cp "$MINECRAFT_FOLDER/versions/1.9.4/1.9.4.jar" jar
-python bin/mapcrafter/src/tools/mapcrafter_textures.py jar/1.9.4.jar bin/mapcrafter/src/data/textures
+if [ -d "$MINECRAFT_FOLDER" ] && [ ! -f "jar/1.9.4.jar" ]; then
+    cp "$MINECRAFT_FOLDER/versions/1.9.4/1.9.4.jar" jar
+fi
+
+if [ -f "jar/1.9.4.jar" ] ; then
+    python bin/mapcrafter/src/tools/mapcrafter_textures.py jar/1.9.4.jar bin/mapcrafter/src/data/textures
+else
+    echo "Minecraft not installed on this computer. Manually download 1.9.4.jar and copy to the jar folder. Then run";
+    echo "'python bin/mapcrafter/src/tools/mapcrafter_textures.py jar/1.9.4.jar bin/mapcrafter/src/data/textures' from";
+    echo "the 'minecraft-realms-map' folder."
+fi
 
 #create render.conf file
 cp render.default render.conf
 
 #create default configuration.conf file
-printf "FTP_SERVER=\nFTP_USERNAME=\nFTP_PASSWORD=\nFTP_PATH=/htdocs\nMINECRAFT_USERNAME=$MINECRAFT_USERNAME\nMINECRAFT_PASSWORD=\nMINECRAFT_PROFILE_ID=$MINECRAFT_PROFILE_ID\nMINECRAFT_CLIENT_ID=$MINECRAFT_CLIENT_ID\nMINECRAFT_VERSION=1.10\nMINECRAFT_WORLD_NUM=1\nHW_THREADS=4" > configuration.conf
+printf "FTP_SERVER=\nFTP_USERNAME=\nFTP_PASSWORD=\nFTP_PATH=/htdocs\nMINECRAFT_USERNAME=$MINECRAFT_USERNAME\nMINECRAFT_PASSWORD=\nMINECRAFT_PROFILE_ID=$MINECRAFT_PROFILE_ID\nMINECRAFT_CLIENT_ID=$MINECRAFT_CLIENT_ID\nMINECRAFT_VERSION=1.10\nMINECRAFT_WORLD_NUM=1\nHW_THREADS=4\nUPLOAD_TYPE=ftp\nS3_URL=s3://bucket-name/" > configuration.conf
